# Remove need for bare-io by using encode_utf8