Fixes #34 

This makes the no_std feature completely obsolete, but it needs to be kept for compatibility (unless a major version bump is done).

There shouldn't be any MSRV concerns as encode_utf8 was stable in 1.15, the MSRV for this crate is at least 1.31 since it uses edition 2018.