@5225225 2022-01-03
Ran the benchmarks (once)

Before:

```
test benches::bench_inlinable_string_from_large                ... bench:          33 ns/iter (+/- 2)
test benches::bench_inlinable_string_from_small                ... bench:           2 ns/iter (+/- 0)
test benches::bench_inlinable_string_push_str_large_onto_empty ... bench:          44 ns/iter (+/- 1)
test benches::bench_inlinable_string_push_str_large_onto_large ... bench:         146 ns/iter (+/- 3)
test benches::bench_inlinable_string_push_str_small_onto_empty ... bench:           2 ns/iter (+/- 0)
test benches::bench_inlinable_string_push_str_small_onto_small ... bench:          10 ns/iter (+/- 0)
test benches::bench_std_string_from_large                      ... bench:          32 ns/iter (+/- 2)
test benches::bench_std_string_from_small                      ... bench:          13 ns/iter (+/- 0)
test benches::bench_std_string_push_str_large_onto_empty       ... bench:          41 ns/iter (+/- 0)
test benches::bench_std_string_push_str_large_onto_large       ... bench:         148 ns/iter (+/- 3)
test benches::bench_std_string_push_str_small_onto_empty       ... bench:          24 ns/iter (+/- 0)
test benches::bench_std_string_push_str_small_onto_small       ... bench:          39 ns/iter (+/- 1)
test inline_string::benches::its_fast                          ... bench:           0 ns/iter (+/- 0)
```

After:

```
test benches::bench_inlinable_string_from_large                ... bench:          33 ns/iter (+/- 1)
test benches::bench_inlinable_string_from_small                ... bench:           2 ns/iter (+/- 0)
test benches::bench_inlinable_string_push_str_large_onto_empty ... bench:          44 ns/iter (+/- 0)
test benches::bench_inlinable_string_push_str_large_onto_large ... bench:         147 ns/iter (+/- 3)
test benches::bench_inlinable_string_push_str_small_onto_empty ... bench:           2 ns/iter (+/- 0)
test benches::bench_inlinable_string_push_str_small_onto_small ... bench:          10 ns/iter (+/- 0)
test benches::bench_std_string_from_large                      ... bench:          31 ns/iter (+/- 1)
test benches::bench_std_string_from_small                      ... bench:          13 ns/iter (+/- 0)
test benches::bench_std_string_push_str_large_onto_empty       ... bench:          36 ns/iter (+/- 2)
test benches::bench_std_string_push_str_large_onto_large       ... bench:         150 ns/iter (+/- 2)
test benches::bench_std_string_push_str_small_onto_empty       ... bench:          18 ns/iter (+/- 1)
test benches::bench_std_string_push_str_small_onto_small       ... bench:          29 ns/iter (+/- 1)
test inline_string::benches::its_fast                          ... bench:           0 ns/iter (+/- 0)
```

I have no idea what that `bench_std_string_push_str_small_onto_small` difference means.