@ArtemGr 2022-01-03
Thanks! Would you bump to `version = "0.1.15"`?