@5225225 2022-01-03
There's also a `miri` failure when running with -Zmiri-tag-raw-pointers, so I'd ideally get that fixed before a release. I'll make a different branch for that (based off this PR because I can't get the code to build without it).