@ArtemGr 2022-03-27
In case somebody's looking for `bare-io`, looks like it has been renamed to `core2`, https://github.com/technocreatives/core2/commit/4072d2a09ba0d656571b1cd2ef6d95aa575c4a69.